Remove expression_impl.rs
