+++ src/sema/expression/assign.rs
@@ -4,7 +4,8 @@ use crate::sema::ast::{Expression, Namespace, RetrieveType, Type};
 use crate::sema::diagnostics::Diagnostics;
 use crate::sema::eval::check_term_for_constant_overflow;
 use crate::sema::expression::integers::get_int_length;
-use crate::sema::expression::{expression, ExprContext, ResolveTo};
+use crate::sema::expression::resolve_expression::expression;
+use crate::sema::expression::{ExprContext, ResolveTo};
 use crate::sema::symtable::Symtable;
 use crate::sema::unused_variable::{assigned_variable, used_variable};
 use crate::sema::Recurse;
