--- src/sema/expression/expression_impl.rs
@@ -1,1203 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-use crate::sema::ast::{ArrayLength, Expression, Namespace, RetrieveType, Type};
-use crate::sema::contracts::is_base;
-use crate::sema::diagnostics::Diagnostics;
-use crate::sema::eval::{eval_const_number, eval_const_rational};
-use crate::sema::expression::compatible_mutability;
-use num_bigint::{BigInt, Sign};
-use num_rational::BigRational;
-use num_traits::ToPrimitive;
-use num_traits::{FromPrimitive, Zero};
-use solang_parser::diagnostics::Diagnostic;
-use solang_parser::pt;
-use solang_parser::pt::CodeLocation;
-use std::cmp::Ordering;
-
-impl Expression {
-    /// Is this expression 0
-    pub(super) fn const_zero(&self, ns: &Namespace) -> bool {
-        if let Ok((_, value)) = eval_const_number(self, ns) {
-            value == BigInt::zero()
-        } else {
-            false
-        }
-    }
-
-    /// Return the type for this expression.
-    pub fn tys(&self) -> Vec<Type> {
-        match self {
-            Expression::Builtin { tys: returns, .. }
-            | Expression::InternalFunctionCall { returns, .. }
-            | Expression::ExternalFunctionCall { returns, .. } => returns.to_vec(),
-            Expression::List { list, .. } => list.iter().map(|e| e.ty()).collect(),
-            Expression::ExternalFunctionCallRaw { .. } => vec![Type::Bool, Type::DynamicBytes],
-            _ => vec![self.ty()],
-        }
-    }
-
-    /// Cast from one type to another, which also automatically derefs any Type::Ref() type.
-    /// if the cast is explicit (e.g. bytes32(bar) then implicit should be set to false.
-    pub(crate) fn cast(
-        &self,
-        loc: &pt::Loc,
-        to: &Type,
-        implicit: bool,
-        ns: &Namespace,
-        diagnostics: &mut Diagnostics,
-    ) -> Result<Expression, ()> {
-        let from = self.ty();
-        if &from == to {
-            return Ok(self.clone());
-        }
-
-        if from == Type::Unresolved || *to == Type::Unresolved {
-            return Ok(self.clone());
-        }
-
-        // First of all, if we have a ref then derefence it
-        if let Type::Ref(r) = &from {
-            return if r.is_fixed_reference_type() {
-                // A struct/fixed array *value* is simply the type, e.g. Type::Struct(_)
-                // An assignable struct value, e.g. member of another struct, is Type::Ref(Type:Struct(_)).
-                // However, the underlying types are identical: simply a pointer.
-                //
-                // So a Type::Ref(Type::Struct(_)) can be cast to Type::Struct(_).
-                //
-                // The Type::Ref(..) just means it can be used as an l-value and assigned
-                // a new value, unlike say, a struct literal.
-                if r.as_ref() == to {
-                    Ok(self.clone())
-                } else {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "conversion from {} to {} not possible",
-                            from.to_string(ns),
-                            to.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                }
-            } else {
-                Expression::Load {
-                    loc: *loc,
-                    ty: r.as_ref().clone(),
-                    expr: Box::new(self.clone()),
-                }
-                .cast(loc, to, implicit, ns, diagnostics)
-            };
-        }
-
-        // If it's a storage reference then load the value. The expr is the storage slot
-        if let Type::StorageRef(_, r) = from {
-            if let Expression::Subscript { array_ty: ty, .. } = self {
-                if ty.is_storage_bytes() {
-                    return Ok(self.clone());
-                }
-            }
-
-            return Expression::StorageLoad {
-                loc: *loc,
-                ty: *r,
-                expr: Box::new(self.clone()),
-            }
-            .cast(loc, to, implicit, ns, diagnostics);
-        }
-
-        // Special case: when converting literal sign can change if it fits
-        match (self, &from, to) {
-            (&Expression::NumberLiteral { ref value, .. }, p, &Type::Uint(to_len))
-                if p.is_primitive() =>
-            {
-                return if value.sign() == Sign::Minus {
-                    if implicit {
-                        diagnostics.push(Diagnostic::cast_error(
-                            *loc,
-                            format!(
-                                "implicit conversion cannot change negative number to '{}'",
-                                to.to_string(ns)
-                            ),
-                        ));
-                        Err(())
-                    } else {
-                        // Convert to little endian so most significant bytes are at the end; that way
-                        // we can simply resize the vector to the right size
-                        let mut bs = value.to_signed_bytes_le();
-
-                        bs.resize(to_len as usize / 8, 0xff);
-                        Ok(Expression::NumberLiteral {
-                            loc: *loc,
-                            ty: Type::Uint(to_len),
-                            value: BigInt::from_bytes_le(Sign::Plus, &bs),
-                        })
-                    }
-                } else if value.bits() >= to_len as u64 {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion would truncate from '{}' to '{}'",
-                            from.to_string(ns),
-                            to.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else {
-                    Ok(Expression::NumberLiteral {
-                        loc: *loc,
-                        ty: Type::Uint(to_len),
-                        value: value.clone(),
-                    })
-                };
-            }
-            (&Expression::NumberLiteral { ref value, .. }, p, &Type::Int(to_len))
-                if p.is_primitive() =>
-            {
-                return if value.bits() >= to_len as u64 {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion would truncate from '{}' to '{}'",
-                            from.to_string(ns),
-                            to.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else {
-                    Ok(Expression::NumberLiteral {
-                        loc: *loc,
-                        ty: Type::Int(to_len),
-                        value: value.clone(),
-                    })
-                };
-            }
-            (&Expression::NumberLiteral { ref value, .. }, p, &Type::Bytes(to_len))
-                if p.is_primitive() =>
-            {
-                // round up the number of bits to bytes
-                let bytes = (value.bits() + 7) / 8;
-                return if value.sign() == Sign::Minus {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "negative number cannot be converted to type '{}'",
-                            to.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else if value.sign() == Sign::Plus && bytes != to_len as u64 {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "number of {} bytes cannot be converted to type '{}'",
-                            bytes,
-                            to.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else {
-                    Ok(Expression::NumberLiteral {
-                        loc: *loc,
-                        ty: Type::Bytes(to_len),
-                        value: value.clone(),
-                    })
-                };
-            }
-            (&Expression::NumberLiteral { ref value, .. }, p, &Type::Address(payable))
-                if p.is_primitive() =>
-            {
-                // note: negative values are allowed
-                return if implicit {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        String::from("implicit conversion from int to address not allowed"),
-                    ));
-                    Err(())
-                } else if value.bits() > ns.address_length as u64 * 8 {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "number larger than possible in {} byte address",
-                            ns.address_length,
-                        ),
-                    ));
-                    Err(())
-                } else {
-                    Ok(Expression::NumberLiteral {
-                        loc: *loc,
-                        ty: Type::Address(payable),
-                        value: value.clone(),
-                    })
-                };
-            }
-            // Literal strings can be implicitly lengthened
-            (&Expression::BytesLiteral { ref value, .. }, p, &Type::Bytes(to_len))
-                if p.is_primitive() =>
-            {
-                return if value.len() > to_len as usize && implicit {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion would truncate from '{}' to '{}'",
-                            from.to_string(ns),
-                            to.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else {
-                    let mut bs = value.to_owned();
-
-                    // Add zero's at the end as needed
-                    bs.resize(to_len as usize, 0);
-
-                    Ok(Expression::BytesLiteral {
-                        loc: *loc,
-                        ty: Type::Bytes(to_len),
-                        value: bs,
-                    })
-                };
-            }
-            (&Expression::BytesLiteral { loc, ref value, .. }, _, &Type::DynamicBytes)
-            | (&Expression::BytesLiteral { loc, ref value, .. }, _, &Type::String) => {
-                return Ok(Expression::AllocDynamicBytes {
-                    loc,
-                    ty: to.clone(),
-                    length: Box::new(Expression::NumberLiteral {
-                        loc,
-                        ty: Type::Uint(32),
-                        value: BigInt::from(value.len()),
-                    }),
-                    init: Some(value.clone()),
-                });
-            }
-            (&Expression::NumberLiteral { ref value, .. }, _, &Type::Rational) => {
-                return Ok(Expression::RationalNumberLiteral {
-                    loc: *loc,
-                    ty: Type::Rational,
-                    value: BigRational::from(value.clone()),
-                });
-            }
-
-            (
-                &Expression::ArrayLiteral { .. },
-                Type::Array(from_ty, from_dims),
-                Type::Array(to_ty, to_dims),
-            ) => {
-                if from_ty == to_ty
-                    && from_dims.len() == to_dims.len()
-                    && from_dims.len() == 1
-                    && matches!(to_dims.last().unwrap(), ArrayLength::Dynamic)
-                {
-                    return Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    });
-                }
-            }
-
-            _ => (),
-        };
-
-        self.cast_types(loc, &from, to, implicit, ns, diagnostics)
-    }
-
-    fn cast_types(
-        &self,
-        loc: &pt::Loc,
-        from: &Type,
-        to: &Type,
-        implicit: bool,
-        ns: &Namespace,
-        diagnostics: &mut Diagnostics,
-    ) -> Result<Expression, ()> {
-        let address_bits = ns.address_length as u16 * 8;
-
-        match (&from, &to) {
-            // Solana builtin AccountMeta struct wants a pointer to an address for the pubkey field,
-            // not an address. For this specific field we have a special Expression::GetRef() which
-            // gets the pointer to an address
-            (Type::Address(_), Type::Ref(to)) if matches!(to.as_ref(), Type::Address(..)) => {
-                Ok(Expression::GetRef {
-                    loc: *loc,
-                    ty: Type::Ref(Box::new(from.clone())),
-                    expr: Box::new(self.clone()),
-                })
-            }
-            (Type::Uint(from_width), Type::Enum(enum_no))
-            | (Type::Int(from_width), Type::Enum(enum_no)) => {
-                if implicit {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion from {} to {} not allowed",
-                            from.to_string(ns),
-                            to.to_string(ns)
-                        ),
-                    ));
-                    return Err(());
-                }
-
-                let enum_ty = &ns.enums[*enum_no];
-
-                // TODO would be help to have current contract to resolve contract constants
-                if let Ok((_, big_number)) = eval_const_number(self, ns) {
-                    if let Some(number) = big_number.to_usize() {
-                        if enum_ty.values.len() > number {
-                            return Ok(Expression::NumberLiteral {
-                                loc: self.loc(),
-                                ty: to.clone(),
-                                value: big_number,
-                            });
-                        }
-                    }
-
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "enum {} has no value with ordinal {}",
-                            to.to_string(ns),
-                            big_number
-                        ),
-                    ));
-                    return Err(());
-                }
-
-                let to_width = enum_ty.ty.bits(ns);
-
-                // TODO needs runtime checks
-                match from_width.cmp(&to_width) {
-                    Ordering::Greater => Ok(Expression::Trunc {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    }),
-                    Ordering::Less => Ok(Expression::ZeroExt {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    }),
-                    Ordering::Equal => Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    }),
-                }
-            }
-            (Type::Enum(enum_no), Type::Uint(to_width))
-            | (Type::Enum(enum_no), Type::Int(to_width)) => {
-                if implicit {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion from {} to {} not allowed",
-                            from.to_string(ns),
-                            to.to_string(ns)
-                        ),
-                    ));
-                    return Err(());
-                }
-                let enum_ty = &ns.enums[*enum_no];
-                let from_width = enum_ty.ty.bits(ns);
-
-                match from_width.cmp(to_width) {
-                    Ordering::Greater => Ok(Expression::Trunc {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    }),
-                    Ordering::Less => Ok(Expression::ZeroExt {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    }),
-                    Ordering::Equal => Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    }),
-                }
-            }
-            (Type::Bytes(n), Type::FunctionSelector) if *n == ns.target.selector_length() => {
-                Ok(Expression::Cast {
-                    loc: *loc,
-                    to: to.clone(),
-                    expr: Box::new(self.clone()),
-                })
-            }
-
-            (Type::Bytes(1), Type::Uint(8)) | (Type::Uint(8), Type::Bytes(1)) => Ok(self.clone()),
-            (Type::Uint(from_len), Type::Uint(to_len)) => match from_len.cmp(to_len) {
-                Ordering::Greater => {
-                    if implicit {
-                        diagnostics.push(Diagnostic::cast_error(
-                            *loc,
-                            format!(
-                                "implicit conversion would truncate from {} to {}",
-                                from.to_string(ns),
-                                to.to_string(ns)
-                            ),
-                        ));
-                        Err(())
-                    } else {
-                        Ok(Expression::Trunc {
-                            loc: *loc,
-                            to: to.clone(),
-                            expr: Box::new(self.clone()),
-                        })
-                    }
-                }
-                Ordering::Less => Ok(Expression::ZeroExt {
-                    loc: *loc,
-                    to: to.clone(),
-                    expr: Box::new(self.clone()),
-                }),
-                Ordering::Equal => Ok(Expression::Cast {
-                    loc: *loc,
-                    to: to.clone(),
-                    expr: Box::new(self.clone()),
-                }),
-            },
-            (Type::Int(from_len), Type::Int(to_len)) => match from_len.cmp(to_len) {
-                Ordering::Greater => {
-                    if implicit {
-                        diagnostics.push(Diagnostic::cast_error(
-                            *loc,
-                            format!(
-                                "implicit conversion would truncate from {} to {}",
-                                from.to_string(ns),
-                                to.to_string(ns)
-                            ),
-                        ));
-                        Err(())
-                    } else {
-                        Ok(Expression::Trunc {
-                            loc: *loc,
-                            to: to.clone(),
-                            expr: Box::new(self.clone()),
-                        })
-                    }
-                }
-                Ordering::Less => Ok(Expression::SignExt {
-                    loc: *loc,
-                    to: to.clone(),
-                    expr: Box::new(self.clone()),
-                }),
-                Ordering::Equal => Ok(Expression::Cast {
-                    loc: *loc,
-                    to: to.clone(),
-                    expr: Box::new(self.clone()),
-                }),
-            },
-            (Type::Uint(from_len), Type::Int(to_len)) if to_len > from_len => {
-                Ok(Expression::ZeroExt {
-                    loc: *loc,
-                    to: to.clone(),
-                    expr: Box::new(self.clone()),
-                })
-            }
-            (Type::Int(from_len), Type::Uint(to_len)) => {
-                if implicit {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion would change sign from {} to {}",
-                            from.to_string(ns),
-                            to.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else if from_len > to_len {
-                    Ok(Expression::Trunc {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    })
-                } else if from_len < to_len {
-                    Ok(Expression::SignExt {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    })
-                } else {
-                    Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    })
-                }
-            }
-            (Type::Uint(from_len), Type::Int(to_len)) => {
-                if implicit {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion would change sign from {} to {}",
-                            from.to_string(ns),
-                            to.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else if from_len > to_len {
-                    Ok(Expression::Trunc {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    })
-                } else if from_len < to_len {
-                    Ok(Expression::ZeroExt {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    })
-                } else {
-                    Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    })
-                }
-            }
-            // Casting value to uint
-            (Type::Value, Type::Uint(to_len)) => {
-                let from_len = ns.value_length * 8;
-                let to_len = *to_len as usize;
-
-                match from_len.cmp(&to_len) {
-                    Ordering::Greater => {
-                        if implicit {
-                            diagnostics.push(Diagnostic::cast_error(
-                                *loc,
-                                format!(
-                                    "implicit conversion would truncate from {} to {}",
-                                    from.to_string(ns),
-                                    to.to_string(ns)
-                                ),
-                            ));
-                            Err(())
-                        } else {
-                            Ok(Expression::Trunc {
-                                loc: *loc,
-                                to: to.clone(),
-                                expr: Box::new(self.clone()),
-                            })
-                        }
-                    }
-                    Ordering::Less => Ok(Expression::SignExt {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    }),
-                    Ordering::Equal => Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    }),
-                }
-            }
-            (Type::Value, Type::Int(to_len)) => {
-                let from_len = ns.value_length * 8;
-                let to_len = *to_len as usize;
-
-                if implicit {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion would change sign from {} to {}",
-                            from.to_string(ns),
-                            to.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else if from_len > to_len {
-                    Ok(Expression::Trunc {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    })
-                } else if from_len < to_len {
-                    Ok(Expression::ZeroExt {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    })
-                } else {
-                    Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    })
-                }
-            }
-            // Casting value to uint
-            (Type::Uint(from_len), Type::Value) => {
-                let from_len = *from_len as usize;
-                let to_len = ns.value_length * 8;
-
-                match from_len.cmp(&to_len) {
-                    Ordering::Greater => {
-                        diagnostics.push(Diagnostic::cast_warning(
-                            *loc,
-                            format!(
-                                "conversion truncates {} to {}, as value is type {} on target {}",
-                                from.to_string(ns),
-                                to.to_string(ns),
-                                Type::Value.to_string(ns),
-                                ns.target
-                            ),
-                        ));
-
-                        Ok(Expression::CheckingTrunc {
-                            loc: *loc,
-                            to: to.clone(),
-                            expr: Box::new(self.clone()),
-                        })
-                    }
-                    Ordering::Less => Ok(Expression::SignExt {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    }),
-                    Ordering::Equal => Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    }),
-                }
-            }
-            // Casting int to address
-            (Type::Uint(from_len), Type::Address(_)) | (Type::Int(from_len), Type::Address(_)) => {
-                if implicit {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion from {} to address not allowed",
-                            from.to_string(ns)
-                        ),
-                    ));
-
-                    Err(())
-                } else {
-                    // cast integer it to integer of the same size of address with sign ext etc
-                    let address_to_int = if from.is_signed_int() {
-                        Type::Int(address_bits)
-                    } else {
-                        Type::Uint(address_bits)
-                    };
-
-                    let expr = match from_len.cmp(&address_bits) {
-                        Ordering::Greater => Expression::Trunc {
-                            loc: *loc,
-                            to: address_to_int,
-                            expr: Box::new(self.clone()),
-                        },
-                        Ordering::Less if from.is_signed_int() => Expression::ZeroExt {
-                            loc: *loc,
-                            to: address_to_int,
-                            expr: Box::new(self.clone()),
-                        },
-                        Ordering::Less => Expression::SignExt {
-                            loc: *loc,
-                            to: address_to_int,
-                            expr: Box::new(self.clone()),
-                        },
-                        Ordering::Equal => self.clone(),
-                    };
-
-                    // Now cast integer to address
-                    Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(expr),
-                    })
-                }
-            }
-            // Casting address to int
-            (Type::Address(_), Type::Uint(to_len)) | (Type::Address(_), Type::Int(to_len)) => {
-                if implicit {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion to {} from {} not allowed",
-                            from.to_string(ns),
-                            to.to_string(ns)
-                        ),
-                    ));
-
-                    Err(())
-                } else {
-                    // first convert address to int/uint
-                    let address_to_int = if to.is_signed_int() {
-                        Type::Int(address_bits)
-                    } else {
-                        Type::Uint(address_bits)
-                    };
-
-                    let expr = Expression::Cast {
-                        loc: *loc,
-                        to: address_to_int,
-                        expr: Box::new(self.clone()),
-                    };
-                    // now resize int to request size with sign extension etc
-                    match to_len.cmp(&address_bits) {
-                        Ordering::Less => Ok(Expression::Trunc {
-                            loc: *loc,
-                            to: to.clone(),
-                            expr: Box::new(expr),
-                        }),
-                        Ordering::Greater if to.is_signed_int() => Ok(Expression::ZeroExt {
-                            loc: *loc,
-                            to: to.clone(),
-                            expr: Box::new(expr),
-                        }),
-                        Ordering::Greater => Ok(Expression::SignExt {
-                            loc: *loc,
-                            to: to.clone(),
-                            expr: Box::new(expr),
-                        }),
-                        Ordering::Equal => Ok(expr),
-                    }
-                }
-            }
-            // Lengthing or shorting a fixed bytes array
-            (Type::Bytes(from_len), Type::Bytes(to_len)) => {
-                if implicit {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion would truncate from {} to {}",
-                            from.to_string(ns),
-                            to.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else if to_len > from_len {
-                    let shift = (to_len - from_len) * 8;
-
-                    Ok(Expression::ShiftLeft {
-                        loc: *loc,
-                        ty: to.clone(),
-                        left: Box::new(Expression::ZeroExt {
-                            loc: self.loc(),
-                            to: to.clone(),
-                            expr: Box::new(self.clone()),
-                        }),
-                        right: Box::new(Expression::NumberLiteral {
-                            loc: *loc,
-                            ty: Type::Uint(*to_len as u16 * 8),
-                            value: BigInt::from_u8(shift).unwrap(),
-                        }),
-                    })
-                } else {
-                    let shift = (from_len - to_len) * 8;
-
-                    Ok(Expression::Trunc {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(Expression::ShiftRight {
-                            loc: self.loc(),
-                            ty: from.clone(),
-                            left: Box::new(self.clone()),
-                            right: Box::new(Expression::NumberLiteral {
-                                loc: self.loc(),
-                                ty: Type::Uint(*from_len as u16 * 8),
-                                value: BigInt::from_u8(shift).unwrap(),
-                            }),
-                            sign: false,
-                        }),
-                    })
-                }
-            }
-            (Type::Rational, Type::Uint(_) | Type::Int(_) | Type::Value) => {
-                match eval_const_rational(self, ns) {
-                    Ok((_, big_number)) => {
-                        if big_number.is_integer() {
-                            return Ok(Expression::Cast {
-                                loc: *loc,
-                                to: to.clone(),
-                                expr: Box::new(self.clone()),
-                            });
-                        }
-
-                        diagnostics.push(Diagnostic::cast_error(
-                            *loc,
-                            format!(
-                                "conversion to {} from {} not allowed",
-                                to.to_string(ns),
-                                from.to_string(ns)
-                            ),
-                        ));
-
-                        Err(())
-                    }
-                    Err(diag) => {
-                        diagnostics.push(diag);
-                        Err(())
-                    }
-                }
-            }
-            (Type::Uint(_) | Type::Int(_) | Type::Value, Type::Rational) => Ok(Expression::Cast {
-                loc: *loc,
-                to: to.clone(),
-                expr: Box::new(self.clone()),
-            }),
-            (Type::Bytes(_), Type::DynamicBytes) | (Type::DynamicBytes, Type::Bytes(_)) => {
-                Ok(Expression::BytesCast {
-                    loc: *loc,
-                    to: to.clone(),
-                    from: from.clone(),
-                    expr: Box::new(self.clone()),
-                })
-            }
-            // Explicit conversion from bytesN to int/uint only allowed with expliciy
-            // cast and if it is the same size (i.e. no conversion required)
-            (Type::Bytes(from_len), Type::Uint(to_len))
-            | (Type::Bytes(from_len), Type::Int(to_len)) => {
-                if implicit {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion to {} from {} not allowed",
-                            to.to_string(ns),
-                            from.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else if *from_len as u16 * 8 != *to_len {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "conversion to {} from {} not allowed",
-                            to.to_string(ns),
-                            from.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else {
-                    Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    })
-                }
-            }
-            // Explicit conversion to bytesN from int/uint only allowed with expliciy
-            // cast and if it is the same size (i.e. no conversion required)
-            (Type::Uint(from_len), Type::Bytes(to_len))
-            | (Type::Int(from_len), Type::Bytes(to_len)) => {
-                if implicit {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion to {} from {} not allowed",
-                            to.to_string(ns),
-                            from.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else if *to_len as u16 * 8 != *from_len {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "conversion to {} from {} not allowed",
-                            to.to_string(ns),
-                            from.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else {
-                    Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    })
-                }
-            }
-            // Explicit conversion from bytesN to address only allowed with expliciy
-            // cast and if it is the same size (i.e. no conversion required)
-            (Type::Bytes(from_len), Type::Address(_)) => {
-                if implicit {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion to {} from {} not allowed",
-                            to.to_string(ns),
-                            from.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else if *from_len as usize != ns.address_length {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "conversion to {} from {} not allowed",
-                            to.to_string(ns),
-                            from.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else {
-                    Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    })
-                }
-            }
-            // Explicit conversion between contract and address is allowed
-            (Type::Address(false), Type::Address(true))
-            | (Type::Address(_), Type::Contract(_))
-            | (Type::Contract(_), Type::Address(_)) => {
-                if implicit {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion to {} from {} not allowed",
-                            to.to_string(ns),
-                            from.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else {
-                    Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    })
-                }
-            }
-            // Conversion between contracts is allowed if it is a base
-            (Type::Contract(contract_no_from), Type::Contract(contract_no_to)) => {
-                if implicit && !is_base(*contract_no_to, *contract_no_from, ns) {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion not allowed since {} is not a base contract of {}",
-                            to.to_string(ns),
-                            from.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else {
-                    Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    })
-                }
-            }
-            // conversion from address payable to address is implicitly allowed (not vice versa)
-            (Type::Address(true), Type::Address(false)) => Ok(Expression::Cast {
-                loc: *loc,
-                to: to.clone(),
-                expr: Box::new(self.clone()),
-            }),
-            // Explicit conversion to bytesN from int/uint only allowed with expliciy
-            // cast and if it is the same size (i.e. no conversion required)
-            (Type::Address(_), Type::Bytes(to_len)) => {
-                if implicit {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "implicit conversion to {} from {} not allowed",
-                            to.to_string(ns),
-                            from.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else if *to_len as usize != ns.address_length {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "conversion to {} from {} not allowed",
-                            to.to_string(ns),
-                            from.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else {
-                    Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    })
-                }
-            }
-            (Type::String, Type::DynamicBytes) | (Type::DynamicBytes, Type::String)
-                if !implicit =>
-            {
-                Ok(Expression::Cast {
-                    loc: *loc,
-                    to: to.clone(),
-                    expr: Box::new(self.clone()),
-                })
-            }
-            // string conversions
-            // (Type::Bytes(_), Type::String) => Ok(Expression::Cast(self.loc(), to.clone(), Box::new(self.clone()))),
-            /*
-            (Type::String, Type::Bytes(to_len)) => {
-                if let Expression::BytesLiteral(_, from_str) = self {
-                    if from_str.len() > to_len as usize {
-                        diagnostics.push(Output::type_error(
-                            self.loc(),
-                            format!(
-                                "string of {} bytes is too long to fit into {}",
-                                from_str.len(),
-                                to.to_string(ns)
-                            ),
-                        ));
-                        return Err(());
-                    }
-                }
-                Ok(Expression::Cast(self.loc(), to.clone(), Box::new(self.clone()))
-            }
-            */
-            (Type::Void, _) => {
-                diagnostics.push(Diagnostic::cast_error(
-                    self.loc(),
-                    "function or method does not return a value".to_string(),
-                ));
-                Err(())
-            }
-            (
-                Type::ExternalFunction {
-                    params: from_params,
-                    mutability: from_mutablity,
-                    returns: from_returns,
-                },
-                Type::ExternalFunction {
-                    params: to_params,
-                    mutability: to_mutablity,
-                    returns: to_returns,
-                },
-            )
-            | (
-                Type::InternalFunction {
-                    params: from_params,
-                    mutability: from_mutablity,
-                    returns: from_returns,
-                },
-                Type::InternalFunction {
-                    params: to_params,
-                    mutability: to_mutablity,
-                    returns: to_returns,
-                },
-            ) => {
-                if from_params != to_params {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "function arguments do not match in conversion from '{}' to '{}'",
-                            to.to_string(ns),
-                            from.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else if from_returns != to_returns {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "function returns do not match in conversion from '{}' to '{}'",
-                            to.to_string(ns),
-                            from.to_string(ns)
-                        ),
-                    ));
-                    Err(())
-                } else if !compatible_mutability(from_mutablity, to_mutablity) {
-                    diagnostics.push(Diagnostic::cast_error(
-                        *loc,
-                        format!(
-                            "function mutability not compatible in conversion from '{}' to '{}'",
-                            from.to_string(ns),
-                            to.to_string(ns),
-                        ),
-                    ));
-                    Err(())
-                } else {
-                    Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: Box::new(self.clone()),
-                    })
-                }
-            }
-            // Match any array with ArrayLength::AnyFixed if is it fixed for that dimension, and the
-            // element type and other dimensions also match
-            (Type::Array(from_elem, from_dim), Type::Array(to_elem, to_dim))
-                if from_elem == to_elem
-                    && from_dim.len() == to_dim.len()
-                    && from_dim.iter().zip(to_dim.iter()).all(|(f, t)| {
-                        f == t || matches!((f, t), (ArrayLength::Fixed(_), ArrayLength::AnyFixed))
-                    }) =>
-            {
-                Ok(self.clone())
-            }
-            (Type::DynamicBytes, Type::Slice(ty)) if ty.as_ref() == &Type::Bytes(1) => {
-                Ok(Expression::Cast {
-                    loc: *loc,
-                    to: to.clone(),
-                    expr: Box::new(self.clone()),
-                })
-            }
-            (Type::FunctionSelector, Type::Bytes(n)) => {
-                let selector_length = ns.target.selector_length();
-                if *n == selector_length {
-                    Ok(Expression::Cast {
-                        loc: *loc,
-                        to: to.clone(),
-                        expr: self.clone().into(),
-                    })
-                } else {
-                    if *n < selector_length {
-                        diagnostics.push(Diagnostic::warning(
-                            *loc,
-                            format!(
-                                "function selector should only be casted to bytes{} or larger",
-                                selector_length
-                            ),
-                        ));
-                    }
-                    self.cast_types(
-                        loc,
-                        &Type::Bytes(selector_length),
-                        to,
-                        implicit,
-                        ns,
-                        diagnostics,
-                    )
-                }
-            }
-            (Type::FunctionSelector, Type::Uint(n) | Type::Int(n)) => {
-                let selector_width = ns.target.selector_length() * 8;
-                if *n < selector_width as u16 {
-                    diagnostics.push(Diagnostic::warning(
-                        *loc,
-                        format!(
-                            "function selector needs an integer of at least {} bits to avoid being truncated",
-                            selector_width
-                        ),
-                    ));
-                }
-                self.cast_types(
-                    loc,
-                    &Type::Bytes(ns.target.selector_length()),
-                    to,
-                    implicit,
-                    ns,
-                    diagnostics,
-                )
-            }
-            _ => {
-                diagnostics.push(Diagnostic::cast_error(
-                    *loc,
-                    format!(
-                        "conversion from {} to {} not possible",
-                        from.to_string(ns),
-                        to.to_string(ns)
-                    ),
-                ));
-                Err(())
-            }
-        }
-    }
-}
