+++ src/sema/expression/literals.rs
@@ -4,8 +4,9 @@ use crate::sema::address::to_hexstr_eip55;
 use crate::sema::ast::{ArrayLength, Expression, Namespace, RetrieveType, StructType, Type};
 use crate::sema::diagnostics::Diagnostics;
 use crate::sema::expression::integers::bigint_to_expression;
+use crate::sema::expression::resolve_expression::expression;
 use crate::sema::expression::strings::unescape;
-use crate::sema::expression::{expression, ExprContext, ResolveTo};
+use crate::sema::expression::{ExprContext, ResolveTo};
 use crate::sema::symtable::Symtable;
 use crate::sema::unused_variable::used_variable;
 use crate::Target;
