+++ src/sema/expression/member_access.rs
@@ -6,7 +6,8 @@ use crate::sema::diagnostics::Diagnostics;
 use crate::sema::expression::constructor::circular_reference;
 use crate::sema::expression::function_call::function_type;
 use crate::sema::expression::integers::bigint_to_expression;
-use crate::sema::expression::{expression, ExprContext, ResolveTo};
+use crate::sema::expression::resolve_expression::expression;
+use crate::sema::expression::{ExprContext, ResolveTo};
 use crate::sema::symtable::Symtable;
 use crate::sema::unused_variable::{assigned_variable, used_variable};
 use num_bigint::{BigInt, Sign};
