+++ src/sema/using.rs
@@ -3,10 +3,11 @@
 use super::{
     ast::{Diagnostic, Expression, Namespace, Note, Type, Using, UsingList},
     diagnostics::Diagnostics,
-    expression::{expression, ExprContext, ResolveTo},
+    expression::{ExprContext, ResolveTo},
     symtable::Symtable,
 };
 use crate::sema::expression::function_call::{function_returns, function_type};
+use crate::sema::expression::resolve_expression::expression;
 use solang_parser::pt;
 use solang_parser::pt::CodeLocation;
 use std::collections::HashSet;
